moved topbar component to components folder
